+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # gpdviz2
 
+2017-08-11
+
+- move web stuff to jvm/src/main/resources
+- preparing for scalaJs
+
 2017-08-10
 
 - tried again to enable scoverage, but got warnings/errors